refactored xrocket code
