+++ xrocket/xrocket/__init__.py
@@ -1,0 +1,1 @@
+from xrocket.encoder import XRocket
